Created integer constants with u suffix as UInt32 (or longer), which is in accordance with the C standard. This removes an erroneous cast in the constant creation and fixes issue #2773

+++ vcc/Core/Parser/Scanner.cs
@@ -1117,7 +1117,7 @@ namespace Microsoft.Research.Vcc.Parsing {
         } else if (ch2 == 'i' || ch2 == 'I') {
           return this.ScanExtendedIntegerSuffix(true);
         }
-        return TypeCode.UInt16;
+        return TypeCode.UInt32;
       } else if (ch == 'l' || ch == 'L') {
         this.endPos++;
         if (ch == 'l') this.HandleError(Error.LowercaseEllSuffix);
